Extend Data base class history tracking to merging histories.
Noticed that cases may arise where multiple Data instances get
merged into one combined object. Data now has a way of tracking
this by combining the histories. The tracking of that is order-
sensitive. Couldn't immediately think of a reason not to!

In the same change, renamed the "input_data" argument to "inputs".