Version 2.0.0-alpha21: Caching and Data classes.
This version finally introduces experimental
caching support into cmlkit.

Caching is implemented on a backend provided by
the Data class, which provides the infrastructure
to track the history of computations applied to a
given initial piece of data. This history is then
used to uniquely identify a Data instance, circumventing
the need to compute (relatively) expensive hashes on
large amounts of data.

The Data class is also responsible for defining
how data is stored to disk, and therefore provides
the heavy lifting for the disk-based cache.

Since this only works if all Components pass instances
of Data to each other, instead of naked ndarrays, this
update introduces quite wide-ranging changes into the
core of cmlkit. It should not break anything, except
custom kernel or KRR implementations, which need to be
transitioned to using the appropriate Data sub-classes.