No longer update hashes on dataset loading
Apparently, we can no longer expect joblib hashes to be stable,
at least for numpy data. So we no longer update it when we load,
instead we just keep it as it was saved in the file. This may
break if stuff gets changed in the file by hand, but that’s
probably an edge case we can safely ignore.